Refactor DNS manager into modular components
